+++ packages/stream_chat_flutter/lib/src/message_widget/message_widget_content.dart
@@ -42,6 +42,7 @@ class MessageWidgetContent extends StatelessWidget {
     required this.hasQuotedMessage,
     required this.hasUrlAttachments,
     required this.hasNonUrlAttachments,
+    required this.hasPoll,
     required this.isOnlyEmoji,
     required this.isGiphy,
     required this.attachmentBuilders,
@@ -137,6 +138,9 @@ class MessageWidgetContent extends StatelessWidget {
 
   /// {@macro hasNonUrlAttachments}
   final bool hasNonUrlAttachments;
+
+  /// {@macro hasPoll}
+  final bool hasPoll;
 
   /// {@macro isOnlyEmoji}
   final bool isOnlyEmoji;
@@ -349,6 +353,7 @@ class MessageWidgetContent extends StatelessWidget {
                                             hasUrlAttachments: hasUrlAttachments,
                                             hasNonUrlAttachments:
                                                 hasNonUrlAttachments,
+                                            hasPoll: hasPoll,
                                             isOnlyEmoji: isOnlyEmoji,
                                             isGiphy: isGiphy,
                                             attachmentBuilders: attachmentBuilders,