+++ packages/stream_chat_flutter/lib/src/theme/stream_chat_theme.dart
@@ -1,4 +1,10 @@
 import 'package:flutter/material.dart' hide TextTheme;
+import 'package:stream_chat_flutter/src/theme/poll_comments_dialog_theme.dart';
+import 'package:stream_chat_flutter/src/theme/poll_interactor_theme.dart';
+import 'package:stream_chat_flutter/src/theme/poll_option_votes_dialog_theme.dart';
+import 'package:stream_chat_flutter/src/theme/poll_options_dialog_theme.dart';
+import 'package:stream_chat_flutter/src/theme/poll_results_dialog_theme.dart';
+import 'package:stream_chat_flutter/src/theme/thread_list_tile_theme.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 /// {@template streamChatTheme}
@@ -56,6 +62,13 @@ class StreamChatThemeData {
     StreamGalleryFooterThemeData? imageFooterTheme,
     StreamMessageListViewThemeData? messageListViewTheme,
     StreamVoiceRecordingThemeData? voiceRecordingTheme,
+    StreamPollCreatorThemeData? pollCreatorTheme,
+    StreamPollInteractorThemeData? pollInteractorTheme,
+    StreamPollOptionsDialogThemeData? pollOptionsDialogTheme,
+    StreamPollResultsDialogThemeData? pollResultsDialogTheme,
+    StreamPollCommentsDialogThemeData? pollCommentsDialogTheme,
+    StreamPollOptionVotesDialogThemeData? pollOptionVotesDialogTheme,
+    StreamThreadListTileThemeData? threadListTileTheme,
   }) {
     brightness ??= colorTheme?.brightness ?? Brightness.light;
     final isDark = brightness == Brightness.dark;
@@ -83,6 +96,13 @@ class StreamChatThemeData {
       galleryFooterTheme: imageFooterTheme,
       messageListViewTheme: messageListViewTheme,
       voiceRecordingTheme: voiceRecordingTheme,
+      pollCreatorTheme: pollCreatorTheme,
+      pollInteractorTheme: pollInteractorTheme,
+      pollOptionsDialogTheme: pollOptionsDialogTheme,
+      pollResultsDialogTheme: pollResultsDialogTheme,
+      pollCommentsDialogTheme: pollCommentsDialogTheme,
+      pollOptionVotesDialogTheme: pollOptionVotesDialogTheme,
+      threadListTileTheme: threadListTileTheme,
     );
 
     return defaultData.merge(customizedData);
@@ -111,6 +131,13 @@ class StreamChatThemeData {
     required this.galleryFooterTheme,
     required this.messageListViewTheme,
     required this.voiceRecordingTheme,
+    required this.pollCreatorTheme,
+    required this.pollInteractorTheme,
+    required this.pollResultsDialogTheme,
+    required this.pollOptionsDialogTheme,
+    required this.pollCommentsDialogTheme,
+    required this.pollOptionVotesDialogTheme,
+    required this.threadListTileTheme,
   });
 
   /// Creates a theme from a Material [Theme]
@@ -160,6 +187,7 @@ class StreamChatThemeData {
         color: const Color(0xff7A7A7A),
       ),
       lastMessageAtStyle: textTheme.footnote.copyWith(
+        // ignore: deprecated_member_use
         color: colorTheme.textHighEmphasis.withOpacity(0.5),
       ),
       indicatorIconSize: 16,
@@ -283,6 +311,202 @@ class StreamChatThemeData {
       voiceRecordingTheme: colorTheme.brightness == Brightness.dark
           ? StreamVoiceRecordingThemeData.dark()
           : StreamVoiceRecordingThemeData.light(),
+      pollCreatorTheme: StreamPollCreatorThemeData(
+        backgroundColor: colorTheme.appBg,
+        appBarBackgroundColor: colorTheme.barsBg,
+        appBarElevation: 1,
+        appBarTitleStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        questionTextFieldFillColor: colorTheme.inputBg,
+        questionHeaderStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        questionTextFieldStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        questionTextFieldErrorStyle: textTheme.footnote.copyWith(
+          color: colorTheme.accentError,
+        ),
+        questionTextFieldBorderRadius: BorderRadius.circular(12),
+        optionsTextFieldFillColor: colorTheme.inputBg,
+        optionsHeaderStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        optionsTextFieldStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        optionsTextFieldErrorStyle: textTheme.footnote.copyWith(
+          color: colorTheme.accentError,
+        ),
+        optionsTextFieldBorderRadius: BorderRadius.circular(12),
+        switchListTileFillColor: colorTheme.inputBg,
+        switchListTileTitleStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        switchListTileErrorStyle: textTheme.footnote.copyWith(
+          color: colorTheme.accentError,
+        ),
+        switchListTileBorderRadius: BorderRadius.circular(12),
+      ),
+      pollInteractorTheme: StreamPollInteractorThemeData(
+        pollTitleStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollSubtitleStyle: textTheme.footnote.copyWith(
+          color: colorTheme.textLowEmphasis,
+        ),
+        pollOptionTextStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionVoteCountTextStyle: textTheme.footnote.copyWith(
+          color: colorTheme.textLowEmphasis,
+        ),
+        pollOptionCheckboxShape: const CircleBorder(),
+        pollOptionCheckboxCheckColor: Colors.white,
+        pollOptionCheckboxActiveColor: colorTheme.accentPrimary,
+        pollOptionCheckboxBorderSide: BorderSide(
+          width: 2,
+          color: colorTheme.disabled,
+        ),
+        pollOptionVotesProgressBarMinHeight: 4,
+        pollOptionVotesProgressBarTrackColor: colorTheme.disabled,
+        pollOptionVotesProgressBarValueColor: colorTheme.accentPrimary,
+        pollOptionVotesProgressBarWinnerColor: colorTheme.accentInfo,
+        pollOptionVotesProgressBarBorderRadius: BorderRadius.circular(4),
+        pollActionButtonStyle: TextButton.styleFrom(
+          textStyle: textTheme.headline,
+          foregroundColor: colorTheme.accentPrimary,
+        ),
+        pollActionDialogTitleStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollActionDialogTextFieldStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollActionDialogTextFieldBorderRadius: BorderRadius.circular(12),
+        pollActionDialogTextFieldFillColor: colorTheme.inputBg,
+      ),
+      pollResultsDialogTheme: StreamPollResultsDialogThemeData(
+        backgroundColor: colorTheme.appBg,
+        appBarElevation: 1,
+        appBarBackgroundColor: colorTheme.barsBg,
+        appBarTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollTitleDecoration: BoxDecoration(
+          color: colorTheme.inputBg,
+          borderRadius: BorderRadius.circular(12),
+        ),
+        pollOptionsDecoration: BoxDecoration(
+          color: colorTheme.inputBg,
+          borderRadius: BorderRadius.circular(12),
+        ),
+        pollOptionsWinnerDecoration: BoxDecoration(
+          color: colorTheme.inputBg,
+          borderRadius: BorderRadius.circular(12),
+        ),
+        pollOptionsTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionsWinnerTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionsVoteCountTextStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionsWinnerVoteCountTextStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionsShowAllVotesButtonStyle: TextButton.styleFrom(
+          textStyle: textTheme.headline,
+          foregroundColor: colorTheme.accentPrimary,
+        ),
+      ),
+      pollOptionsDialogTheme: StreamPollOptionsDialogThemeData(
+        backgroundColor: colorTheme.appBg,
+        appBarElevation: 1,
+        appBarBackgroundColor: colorTheme.barsBg,
+        appBarTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollTitleDecoration: BoxDecoration(
+          color: colorTheme.inputBg,
+          borderRadius: BorderRadius.circular(12),
+        ),
+        pollOptionsListViewDecoration: BoxDecoration(
+          color: colorTheme.inputBg,
+          borderRadius: BorderRadius.circular(12),
+        ),
+      ),
+      pollCommentsDialogTheme: StreamPollCommentsDialogThemeData(
+        backgroundColor: colorTheme.appBg,
+        appBarElevation: 1,
+        appBarBackgroundColor: colorTheme.barsBg,
+        appBarTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollCommentItemBackgroundColor: colorTheme.inputBg,
+        pollCommentItemBorderRadius: BorderRadius.circular(12),
+        updateYourCommentButtonStyle: TextButton.styleFrom(
+          textStyle: textTheme.headlineBold,
+          foregroundColor: colorTheme.accentPrimary,
+          backgroundColor: colorTheme.inputBg,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+          padding: const EdgeInsets.symmetric(
+            vertical: 18,
+            horizontal: 16,
+          ),
+        ),
+      ),
+      pollOptionVotesDialogTheme: StreamPollOptionVotesDialogThemeData(
+        backgroundColor: colorTheme.appBg,
+        appBarElevation: 1,
+        appBarBackgroundColor: colorTheme.barsBg,
+        appBarTitleTextStyle: textTheme.headlineBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionVoteCountTextStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionWinnerVoteCountTextStyle: textTheme.headline.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        pollOptionVoteItemBackgroundColor: colorTheme.inputBg,
+        pollOptionVoteItemBorderRadius: BorderRadius.circular(12),
+      ),
+      threadListTileTheme: StreamThreadListTileThemeData(
+        backgroundColor: colorTheme.barsBg,
+        padding: const EdgeInsets.symmetric(vertical: 14, horizontal: 8),
+        threadUnreadMessageCountStyle: textTheme.footnoteBold.copyWith(
+          color: Colors.white,
+        ),
+        threadUnreadMessageCountBackgroundColor:
+            channelPreviewTheme.unreadCounterColor,
+        threadChannelNameStyle: textTheme.bodyBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        threadReplyToMessageStyle: textTheme.footnote.copyWith(
+          color: colorTheme.textLowEmphasis,
+        ),
+        threadLatestReplyTimestampStyle: textTheme.footnote.copyWith(
+          color: colorTheme.textLowEmphasis,
+        ),
+        threadLatestReplyUsernameStyle: textTheme.bodyBold.copyWith(
+          color: colorTheme.textHighEmphasis,
+        ),
+        threadLatestReplyMessageStyle: textTheme.body.copyWith(
+          color: colorTheme.textLowEmphasis,
+        ),
+      ),
     );
   }
 
@@ -327,6 +551,27 @@ class StreamChatThemeData {
   /// Theme configuration for the [StreamVoiceRecordingListPLayer] widget.
   final StreamVoiceRecordingThemeData voiceRecordingTheme;
 
+  /// Theme configuration for the [StreamPollCreatorWidget] widget.
+  final StreamPollCreatorThemeData pollCreatorTheme;
+
+  /// Theme configuration for the [StreamPollInteractor] widget.
+  final StreamPollInteractorThemeData pollInteractorTheme;
+
+  /// Theme configuration for the [StreamPollResultsDialog] widget.
+  final StreamPollResultsDialogThemeData pollResultsDialogTheme;
+
+  /// Theme configuration for the [StreamPollOptionsDialog] widget.
+  final StreamPollOptionsDialogThemeData pollOptionsDialogTheme;
+
+  /// Theme configuration for the [StreamPollCommentsDialog] widget.
+  final StreamPollCommentsDialogThemeData pollCommentsDialogTheme;
+
+  /// Theme configuration for the [StreamPollOptionVotesDialog] widget.
+  final StreamPollOptionVotesDialogThemeData pollOptionVotesDialogTheme;
+
+  /// Theme configuration for the [StreamThreadListTile] widget.
+  final StreamThreadListTileThemeData threadListTileTheme;
+
   /// Creates a copy of [StreamChatThemeData] with specified attributes
   /// overridden.
   StreamChatThemeData copyWith({
@@ -347,6 +592,13 @@ class StreamChatThemeData {
     StreamGalleryFooterThemeData? galleryFooterTheme,
     StreamMessageListViewThemeData? messageListViewTheme,
     StreamVoiceRecordingThemeData? voiceRecordingTheme,
+    StreamPollCreatorThemeData? pollCreatorTheme,
+    StreamPollInteractorThemeData? pollInteractorTheme,
+    StreamPollResultsDialogThemeData? pollResultsDialogTheme,
+    StreamPollOptionsDialogThemeData? pollOptionsDialogTheme,
+    StreamPollCommentsDialogThemeData? pollCommentsDialogTheme,
+    StreamPollOptionVotesDialogThemeData? pollOptionVotesDialogTheme,
+    StreamThreadListTileThemeData? threadListTileTheme,
   }) =>
       StreamChatThemeData.raw(
         channelListHeaderTheme:
@@ -364,6 +616,17 @@ class StreamChatThemeData {
         galleryFooterTheme: galleryFooterTheme ?? this.galleryFooterTheme,
         messageListViewTheme: messageListViewTheme ?? this.messageListViewTheme,
         voiceRecordingTheme: voiceRecordingTheme ?? this.voiceRecordingTheme,
+        pollCreatorTheme: pollCreatorTheme ?? this.pollCreatorTheme,
+        pollInteractorTheme: pollInteractorTheme ?? this.pollInteractorTheme,
+        pollResultsDialogTheme:
+            pollResultsDialogTheme ?? this.pollResultsDialogTheme,
+        pollOptionsDialogTheme:
+            pollOptionsDialogTheme ?? this.pollOptionsDialogTheme,
+        pollCommentsDialogTheme:
+            pollCommentsDialogTheme ?? this.pollCommentsDialogTheme,
+        pollOptionVotesDialogTheme:
+            pollOptionVotesDialogTheme ?? this.pollOptionVotesDialogTheme,
+        threadListTileTheme: threadListTileTheme ?? this.threadListTileTheme,
       );
 
   /// Merge themes
@@ -385,6 +648,17 @@ class StreamChatThemeData {
       messageListViewTheme:
           messageListViewTheme.merge(other.messageListViewTheme),
       voiceRecordingTheme: voiceRecordingTheme.merge(other.voiceRecordingTheme),
+      pollCreatorTheme: pollCreatorTheme.merge(other.pollCreatorTheme),
+      pollInteractorTheme: pollInteractorTheme.merge(other.pollInteractorTheme),
+      pollResultsDialogTheme:
+          pollResultsDialogTheme.merge(other.pollResultsDialogTheme),
+      pollOptionsDialogTheme:
+          pollOptionsDialogTheme.merge(other.pollOptionsDialogTheme),
+      pollCommentsDialogTheme:
+          pollCommentsDialogTheme.merge(other.pollCommentsDialogTheme),
+      pollOptionVotesDialogTheme:
+          pollOptionVotesDialogTheme.merge(other.pollOptionVotesDialogTheme),
+      threadListTileTheme: threadListTileTheme.merge(other.threadListTileTheme),
     );
   }
 }